+++ app/src/main/java/kr/co/nottodo/view/NoChangeAnimationRecyclerView.kt
@@ -1,4 +1,4 @@
-package kr.co.nottodo.view.monthcalendar
+package kr.co.nottodo.view
 
 import android.content.Context
 import android.util.AttributeSet
