+++ app/src/main/java/kr/co/nottodo/view/monthlycalendar/adapter/NotToDoDayAdapter.kt
@@ -1,4 +1,4 @@
-package kr.co.nottodo.view.monthcalendar.common.adapter
+package kr.co.nottodo.view.monthlycalendar.adapter
 
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
@@ -8,11 +8,11 @@ import androidx.recyclerview.widget.RecyclerView
 import kr.co.nottodo.R
 import kr.co.nottodo.databinding.ViewNotToDoCalendarDayBinding
 import kr.co.nottodo.databinding.ViewNotToDoCalendarEmptyBinding
-import kr.co.nottodo.view.monthcalendar.common.NotToDoCalendarDay
-import kr.co.nottodo.view.monthcalendar.common.CalendarType
-import kr.co.nottodo.view.monthcalendar.common.viewholder.NotToDoCalendarViewHolder
-import kr.co.nottodo.view.monthcalendar.common.viewholder.NotToDoDayViewHolder
-import kr.co.nottodo.view.monthcalendar.common.viewholder.NotToDoEmptyViewHolder
+import kr.co.nottodo.view.monthlycalendar.model.NotToDoCalendarDay
+import kr.co.nottodo.view.monthlycalendar.model.CalendarType
+import kr.co.nottodo.view.monthlycalendar.viewholder.NotToDoCalendarViewHolder
+import kr.co.nottodo.view.monthlycalendar.viewholder.NotToDoDayViewHolder
+import kr.co.nottodo.view.monthlycalendar.viewholder.NotToDoEmptyViewHolder
 
 class NotToDoDayAdapter: RecyclerView.Adapter<NotToDoCalendarViewHolder>() {
