+++ app/src/main/java/kr/co/nottodo/view/monthlycalendar/model/NotToDoCalendarEntity.kt
@@ -1,4 +1,4 @@
-package kr.co.nottodo.view.monthcalendar.common
+package kr.co.nottodo.view.monthlycalendar.model
 
 import java.util.*
