+++ app/src/main/java/kr/co/nottodo/view/monthlycalendar/util/CalendarUtil.kt
@@ -1,4 +1,4 @@
-package kr.co.nottodo.view.monthcalendar.common.util
+package kr.co.nottodo.view.monthlycalendar.util
 
 import java.util.*
 import java.util.Calendar.*
