+++ app/src/main/java/kr/co/nottodo/view/monthlycalendar/viewholder/NotToDoCalendarViewHolder.kt
@@ -1,8 +1,8 @@
-package kr.co.nottodo.view.monthcalendar.common.viewholder
+package kr.co.nottodo.view.monthlycalendar.viewholder
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import kr.co.nottodo.view.monthcalendar.common.NotToDoCalendarDay
+import kr.co.nottodo.view.monthlycalendar.model.NotToDoCalendarDay
 
 abstract class NotToDoCalendarViewHolder(view: View) : RecyclerView.ViewHolder(view) {
     abstract fun onBind(data: NotToDoCalendarDay)
