+++ app/src/main/java/kr/co/nottodo/view/weeklycalendar/NotToDoWeeklyAdapter.kt
@@ -1,0 +1,4 @@
+package kr.co.nottodo.view.weeklycalendar
+
+class NotToDoWeeklyAdapter {
+}
