+++ app/src/main/java/kr/co/nottodo/view/weeklycalendar/NotToDoWeeklyCalendar.kt
@@ -1,0 +1,13 @@
+package kr.co.nottodo.view.weeklycalendar
+
+import android.content.Context
+import android.util.AttributeSet
+import androidx.recyclerview.widget.RecyclerView
+
+class NotToDoWeeklyCalendar @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0
+): RecyclerView(context,attrs,defStyleAttr) {
+
+}
