+++ app/src/main/res/layout/view_not_to_do_calendar_empty.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="empty_day_item"
-            type="kr.co.nottodo.view.monthcalendar.common.NotToDoCalendarDay.Empty" />
+            type="kr.co.nottodo.view.monthlycalendar.model.NotToDoCalendarDay.Empty" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
