+++ app/src/main/res/layout/view_not_to_do_calendar_month.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="month_item"
-            type="kr.co.nottodo.view.monthcalendar.common.NotToDoCalendarMonth" />
+            type="kr.co.nottodo.view.monthlycalendar.model.NotToDoCalendarMonth" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
