+++ app/src/main/res/layout/view_not_to_do_weekly_calendar.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <include
+        android:id="@+id/view_calendar_week_description"
+        layout="@layout/view_not_to_do_calendar_week_description"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:layout_constraintStart_toStartOf="@id/view_calendar_week_description"
+        app:layout_constraintEnd_toEndOf="@id/view_calendar_week_description"
+        app:layout_constraintTop_toBottomOf="@id/view_calendar_week_description"/>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
